Create the first version of the interactive view.
This produces an HTML page with largely the same info and
style as the calendar view. The transparency rules are not
exactly the same and the view has some graphical differences,
but nothing large.

Immediately this has two big advantages over the picture :
It will size itself to the screen (so it's always as big as
possible), and as there is a minimum width per day it will
scroll horizontally when there are many days. These alone
make this representation already superior to the old one.
There also is a tooltip shown when hovering over any block
of time, which the image has no chance of doing. Right now
this only displays the activity in its final state so it's
not too useful yet.

At this time there are a few small bugs : the horizontal
lines for the ruler for time stop at one screenful when
scrolling horizontally. In some cases apparently the grid
showing days has one too many days on the right.

Future improvements will see the tooltip massively
improved in usefulness, but this will require some infra
changes as the model doesn't contain enough information
yet.