Store the entire hierarchy of categories for all activities
Each category has a link to its parent (or nil if none). This
makes it possible to display the entire list of embedded
categories on any activity.

This patch also changes how colors are specified in such a way
that the color displayed for an activity will always be the one
for the nearest specified parent color. It was the obvious thing
to display, and has the large advantage that the rules files can
specify colors for embedded categories and the outputs will use
the most specific ones, which is really useful.

This makes it apparent that the current design that separates
the activity (as a String) and the category (as a hierarchy of
strings) is no longer great. It would be simpler to collapse
them and just say that each activity is a leaf category with
a parent link to its own specified category.