Support split activities
This breaks most modes probably if they weren't broken already,
but it adds support in the model and the interactive visualization
for split activities, where you're doing multiple things at the
same time and account for both of them.

Mode occupation and mode stack were probably hopelessly broken
anyway, but this completely breaks mode count (I won't even dare
try and run it to see how broken it is). This eventually won't
matter when the interactive visualization shows the counts in
a much better way, and then I can remove the old mode.