Commit old cleanups
I don't even remember when I wrote this patch. It's at least one year
old.

It looks like this is rationalizing data structures in util (switching from
an informal array to a structured record for dates) and adding
support for percentages in rule parsing for forward compatibility
with log2 (support for expressions like "40% 60%" in rule files).